+++ Flipdish.Recruiting.WebhookReceiver/Flipdish.Recruiting.WebhookReceiver.csproj
@@ -6,6 +6,11 @@
     <AssemblyName>Flipdish.Recruiting.WebhookReceiver</AssemblyName>
   </PropertyGroup>
   <ItemGroup>
+    <Compile Remove="Properties\**" />
+    <EmbeddedResource Remove="Properties\**" />
+    <None Remove="Properties\**" />
+  </ItemGroup>
+  <ItemGroup>
     <PackageReference Include="DotLiquid" Version="2.0.366" />
     <PackageReference Include="Microsoft.Azure.Functions.Extensions" Version="1.1.0" />
     <PackageReference Include="Microsoft.NET.Sdk.Functions" Version="3.0.7" />
@@ -40,9 +45,6 @@
     <None Update="TestWebhooks\1.json">
       <CopyToOutputDirectory>Always</CopyToOutputDirectory>
     </None>
-  </ItemGroup>
-  <ItemGroup>
-    <Folder Include="Properties\" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\Flipdish.Recruiting.Domain\Flipdish.Recruiting.Domain.csproj" />
